+++ android/src/oldarch/com/appodeal/rnappodeal/RNAppodealModule.java
@@ -7,6 +7,7 @@ import com.facebook.react.bridge.ReactApplicationContext;
 import com.facebook.react.bridge.ReactContextBaseJavaModule;
 import com.facebook.react.bridge.ReactMethod;
 import com.facebook.react.bridge.ReadableMap;
+import com.facebook.react.bridge.WritableArray;
 import com.facebook.react.bridge.WritableMap;
 
 public class RNAppodealModule extends ReactContextBaseJavaModule {
@@ -237,6 +238,31 @@ public class RNAppodealModule extends ReactContextBaseJavaModule {
     @ReactMethod
     public void trackEvent(String name, ReadableMap parameters) {
         moduleImplementation.trackEvent(name, parameters);
+    }
+
+    @ReactMethod(isBlockingSynchronousMethod = true)
+    public WritableArray getNativeAds(double count) {
+        return moduleImplementation.getNativeAds(count);
+    }
+
+    @ReactMethod(isBlockingSynchronousMethod = true)
+    public double getAvailableNativeAdsCount() {
+        return moduleImplementation.getAvailableNativeAdsCount();
+    }
+
+    @ReactMethod
+    public void destroyNativeAd(String adId) {
+        moduleImplementation.destroyNativeAd(adId);
+    }
+
+    @ReactMethod
+    public void cacheNativeAds(double count) {
+        moduleImplementation.cacheNativeAds(count);
+    }
+
+    @ReactMethod
+    public void setPreferredNativeContentType(String type) {
+        moduleImplementation.setPreferredNativeContentType(type);
     }
 
     @ReactMethod